rename installer to checker
