+++ scivision/io/__init__.py
@@ -1,3 +1,3 @@
 from .reader import load_pretrained_model, load_dataset, _parse_url, _get_model_configs
-from .installer import check_package
+from .checker import check_package
 from .wrapper import PretrainedModel, Datasource
