+++ scivision/io/checker.py
